+++ easy/1228_missing_number_in_arithmetic_progression/solution.cpp
@@ -3,13 +3,13 @@
 class Solution {
 public:
   int missingNumber(std::vector<int> &arr) {
-    const int arrLen = arr.size();
+    const size_t arrLen = arr.size();
     const int step = (arr[arrLen - 1] - arr[0]) / arrLen;
 
     if (step == 0)
       return arr[0];
 
-    for (int i = 0, expected = arr[0]; expected != arr[arrLen - 1];
+    for (size_t i = 0, expected = arr[0]; expected != arr[arrLen - 1];
          ++i, expected += step) {
       if (arr[i] != expected)
         return expected;
