+++ easy/2073_time_needed_to_buy_tickets/solution.cpp
@@ -5,7 +5,7 @@ public:
   int timeRequiredToBuy(std::vector<int> &tickets, int k) {
     int time = 0;
 
-    for (int i = 0; i < tickets.size(); ++i) {
+    for (size_t i = 0; i < tickets.size(); ++i) {
       time += std::min(tickets[i], tickets[k] - (i > k));
     }
