+++ easy/3379_transformed_array/solution.cpp
@@ -3,11 +3,11 @@
 class Solution {
 public:
   std::vector<int> constructTransformedArray(std::vector<int> &nums) {
-    const int numsLen = nums.size();
+    const size_t numsLen = nums.size();
 
     std::vector<int> transformedNums(numsLen, 0);
 
-    for (int i = 0; i < numsLen; ++i) {
+    for (size_t i = 0; i < numsLen; ++i) {
       int ii = i + nums[i];
       while (ii < 0)
         ii += numsLen;
