+++ easy/3861_minimum_capacity_box/solution.cpp
@@ -5,7 +5,7 @@ public:
   int minimumIndex(std::vector<int> &capacity, int itemSize) {
     int minI = 0;
 
-    for (int i = 0; i < capacity.size(); ++i) {
+    for (size_t i = 0; i < capacity.size(); ++i) {
       if (capacity[i] >= itemSize &&
           (capacity[minI] < itemSize || capacity[i] < capacity[minI]))
         minI = i;
